Add keyword search to reports in history
I implemented a keyword search using the gem 'pg_search'. It provides a very straightforward way of taking advantage of active record scopes in order to make SQL queries. I built the search scope in models/report.rb and you can see that it is pretty simple and extendable. Adding or removing columns and associations should be very easy. The gem offers some easily implementable features like trigrams and double metaphones that could be very useful in the future when lots of reports are in the database.