allow keyword search to filter all reports
Added functionality to search reports of all statuses, not only those completed. This is in order to provide the use case where a dispatcher can use keyword search in history and return reports that are currently in unassigned and pending.

Something to note is that report and dispatch states are different and should factor in to the expectations for search results. Reports that are filed under "unassigned" and "pending" in reports/index.html are only differentiated by their dispatch state; as far as reports are concerned, they are both "pending". For this PR we are searching against report =>status, so a search for "pending" will return reports that have a dispatch of both "unassigned" and "pending".